requirements.txt
# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# High Energy Physics specific
uproot>=4.0.0
awkward>=1.0.0
h5py>=3.0.0

# Utilities
tqdm>=4.60.0
pyyaml>=5.4.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0
